app-engine/api/requirements.txt
Flask==1.1.1
google-cloud-datastore==1.9.0
google-cloud-secret-manager==0.2.0
pytz==2019.2
sentry-sdk==0.14.2